requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
discord.py==2.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
multidict==6.0.3
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.21.0
rfc3986==1.5.0
sniffio==1.3.0
yarl==1.8.2
youtube-dl==2021.12.17
youtube-search-python==1.6.6
